# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Stacklab API
  version: 0.1.0
  description: Compose-first homelab control panel API. Every HTTP response includes `X-Request-ID`; the same optional request header is accepted when it matches the safe identifier format.
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /config/workspace/tree:
    get:
      tags: [Config]
      operationId: getConfigWorkspaceTree
      parameters:
        - in: query
          name: path
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Directory-scoped config workspace listing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigTreeResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '400':
          description: Invalid path or path is not a directory.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /config/workspace/file:
    get:
      tags: [Config]
      operationId: getConfigWorkspaceFile
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Config workspace file metadata and optional content.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigFileResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '400':
          description: Invalid path or path is not a file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '413':
          $ref: '#/components/responses/ErrorContentTooLarge'
    put:
      tags: [Config]
      operationId: saveConfigWorkspaceFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigFileSaveRequest'
      responses:
        '200':
          description: Config workspace file saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigFileSaveResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '400':
          description: Invalid path or parent directory shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          description: Binary files are not editable in-browser.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorContentTooLarge'
    delete:
      tags: [Config]
      operationId: deleteConfigWorkspaceFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigFileDeleteRequest'
      responses:
        '200':
          description: Config workspace file deleted without committing or pushing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigFileDeleteResponse'
        '400':
          description: Invalid path, missing timestamp, or target is not a regular file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          description: Cross-origin request rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          description: File changed since it was loaded or deletion is blocked by permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /config/workspace/repair-permissions:
    post:
      tags: [Config]
      operationId: repairConfigWorkspacePermissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigRepairPermissionsRequest'
      responses:
        '200':
          description: Config workspace ownership and owner access repaired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigRepairPermissionsResponse'
        '400':
          description: Path escapes the managed config workspace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '501':
          description: Workspace permission repair is not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /git/workspace/status:
    get:
      tags: [Git]
      operationId: getGitWorkspaceStatus
      responses:
        '200':
          description: Read-only Git status for the managed Stacklab workspace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitWorkspaceStatusResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '413':
          $ref: '#/components/responses/ErrorContentTooLarge'
  /git/workspace/diff:
    get:
      tags: [Git]
      operationId: getGitWorkspaceDiff
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unified diff for one changed managed file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitDiffResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '400':
          description: Invalid path or path outside managed roots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '413':
          $ref: '#/components/responses/ErrorContentTooLarge'
        '503':
          description: Git workspace is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /git/workspace/file:
    delete:
      tags: [Git]
      operationId: deleteGitWorkspaceFile
      description: Delete one changed regular file under config/ or a stack directory. Active stack compose.yaml and .env files, directories, and symbolic links are protected. Does not stage, commit, or push.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitFileDeleteRequest'
      responses:
        '200':
          description: The selected workspace file was deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitFileDeleteResponse'
        '400':
          description: Invalid path, missing timestamp, or a protected target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          description: Cross-origin request rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          description: The file changed, permissions block deletion, or a Git mutation is in progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Git workspace is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /git/workspace/commit:
    post:
      tags: [Git]
      operationId: commitGitWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitCommitRequest'
      responses:
        '200':
          description: Local commit created for selected managed files.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitCommitResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '400':
          description: Invalid commit request or invalid path selection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          description: Commit cannot proceed due to conflicts, permissions, or no staged changes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Git workspace is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /git/workspace/push:
    post:
      tags: [Git]
      operationId: pushGitWorkspace
      responses:
        '200':
          description: Current branch push result for configured upstream.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitPushResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '409':
          description: Upstream is missing, a required file is inaccessible, or remote rejected the push.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Remote authentication, repository access, or SSH host verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Git workspace is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
      description: Dedicated metrics token loaded from STACKLAB_METRICS_TOKEN_FILE. Grants access only to GET /metrics.
    sessionCookie:
      type: apiKey
      in: cookie
      name: stacklab_session
      description: Opaque HTTP-only session cookie. Successful authenticated responses refresh it to the immutable absolute session deadline; the server-side idle lease remains authoritative.
...
  responses:
    ErrorUnauthorized:
      description: Authentication required or invalid session.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ErrorNotFound:
      description: Resource not found.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ErrorContentTooLarge:
      description: File content or generated command output exceeds its bounded processing limit.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ConfigEntryType:
      type: string
      enum: [directory, text_file, binary_file, unknown_file]
    ConfigTreeResponse:
      type: object
      properties:
        workspace_root:
          type: string
        current_path:
          type: string
        parent_path:
          type: [string, 'null']
        items:
          type: array
          items:
            $ref: '#/components/schemas/ConfigTreeEntry'
      required: [workspace_root, current_path, parent_path, items]
...
    ConfigFileResponse:
      type: object
      properties:
        path:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/ConfigEntryType'
        stack_id:
          type: [string, 'null']
        content:
          type: [string, 'null']
        encoding:
          type: [string, 'null']
        size_bytes:
          type: integer
        modified_at:
          type: string
          format: date-time
        git_ignored:
          type: boolean
        readable:
          type: boolean
        writable:
          type: boolean
        blocked_reason:
          type: [string, 'null']
        permissions:
          $ref: '#/components/schemas/FilePermissions'
        repair_capability:
          $ref: '#/components/schemas/WorkspaceRepairCapability'
      required: [path, name, type, stack_id, content, encoding, size_bytes, modified_at, git_ignored, readable, writable, blocked_reason, permissions, repair_capability]
    ConfigFileSaveRequest:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        create_parent_directories:
          type: boolean
        expected_modified_at:
          type: [string, 'null']
          format: date-time
      required: [path, content, create_parent_directories]
    ConfigFileSaveResponse:
      type: object
      properties:
        saved:
          type: boolean
        path:
          type: string
        modified_at:
          type: string
          format: date-time
        audit_action:
          type: string
      required: [saved, path, modified_at, audit_action]
    ConfigFileDeleteRequest:
      type: object
      properties:
        path:
          type: string
          minLength: 1
        expected_modified_at:
          type: string
          format: date-time
      required: [path, expected_modified_at]
    ConfigFileDeleteResponse:
      type: object
      properties:
        deleted:
          type: boolean
        path:
          type: string
        audit_action:
          type: string
      required: [deleted, path, audit_action]
...
    ConfigRepairPermissionsRequest:
      type: object
      properties:
        path:
          type: string
        recursive:
          type: boolean
          default: false
      required: [path]
    ConfigRepairPermissionsResponse:
      type: object
      properties:
        repaired:
          type: boolean
        path:
          type: string
        recursive:
          type: boolean
        changed_items:
          type: integer
        warnings:
          type: array
          items:
            type: string
        target_permissions_before:
          $ref: '#/components/schemas/FilePermissions'
        target_permissions_after:
          $ref: '#/components/schemas/FilePermissions'
        audit_action:
          type: string
        repair_capability:
          $ref: '#/components/schemas/WorkspaceRepairCapability'
      required: [repaired, path, recursive, changed_items, target_permissions_before, target_permissions_after, audit_action, repair_capability]
...
    GitWorkspaceStatusResponse:
      type: object
      properties:
        available:
          type: boolean
        repo_root:
          type: string
        managed_roots:
          type: array
          items:
            type: string
        branch:
          type: string
        head_commit:
          type: string
        has_upstream:
          type: boolean
        upstream_name:
          type: string
        ahead_count:
          type: integer
        behind_count:
          type: integer
        clean:
          type: boolean
        reason:
          type: string
          enum: [git_not_installed, not_a_git_repository]
        items:
          type: array
          items:
            $ref: '#/components/schemas/GitStatusItem'
      required: [available, repo_root, managed_roots, clean]
...
    GitDiffResponse:
      type: object
      properties:
        delete_allowed:
          type: boolean
          description: The path identifies a deletable regular file; actual parent directory permissions are checked on deletion.
        modified_at:
          type: string
          format: date-time
          description: Required as expected_modified_at when deleting the reviewed file.
        repair_capability:
          $ref: '#/components/schemas/WorkspaceRepairCapability'
        available:
          type: boolean
        path:
          type: string
        scope:
          $ref: '#/components/schemas/GitScope'
        stack_id:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/GitFileStatus'
        old_path:
          type: [string, 'null']
        permissions:
          $ref: '#/components/schemas/FilePermissions'
        diff_available:
          type: boolean
        blocked_reason:
          type: [string, 'null']
          enum: [not_readable, null]
        is_binary:
          type: boolean
        diff:
          type: [string, 'null']
        truncated:
          type: boolean
      required: [available, path, scope, stack_id, status, old_path, diff_available, blocked_reason, is_binary, diff, truncated]
...
    GitFileDeleteRequest:
      type: object
      properties:
        path:
          type: string
          description: Full managed path, such as stacks/samba/compose.yaml.bak or config/samba/obsolete.conf.
        expected_modified_at:
          type: string
          format: date-time
      required: [path, expected_modified_at]
...
    GitCommitRequest:
      type: object
      properties:
        message:
          type: string
        paths:
          type: array
          items:
            type: string
      required: [message, paths]
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required: [error]
...